Gato/Perro
Ahora el jugador no puede moverse si muere, y el gato/perro puede quitarle vida y detectarlo, pero no logro que lo persiga :c

+++ Assets/Scripts/Car/IDamageable.cs
@@ -4,5 +4,5 @@ using System.Collections.Generic;
 
 public interface IDamageable 
 {
-    void TakeDamage(int amount); //dont care about what kind of object deal damage, just apply it
+    void TakeDamage(float amount); //dont care about what kind of object deal damage, just apply it
 }
